backend/services/ledger-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
pydantic-settings
python-dateutil
